backend/requirements.txt
haystack-ai==2.10.3
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
elasticsearch==8.12.1
sentence-transformers==3.4.1
python-dotenv>=1.0.0
pydantic>=2.5.2
pydantic-settings>=2.0.0
pypdf>=3.0.0
markdown-it-py>=3.0.0
mdit_plain>=1.0.1
elasticsearch-haystack==2.1.0
qdrant-haystack==8.0.0
qdrant-client==1.13.2
ollama-haystack==2.3.0
nltk>=3.9.1
mysql-connector-python>=8.0.33
